feat(rebalance): 소수점 거래 — 포트별 설정, 기본은 온주 (v23b)
증권사·계좌마다 소수점 매매 가능 여부가 다르므로 앱 전역 설정이 아니라
포트폴리오별 설정으로 두고, 기본값은 지금까지와 같은 온주(1주 단위)로 뒀다.
예전 저장 데이터에는 이 키가 없으므로 자동으로 온주 거래로 읽힌다.

계산:
- RebalanceItemResult.newShares / delta 를 int → double 로.
  PortfolioItem.shares 와 StockTransaction.quantity 는 이미 double이라
  모델 저장 형식은 그대로다.
- 소수점 거래에서는 목표 금액을 그대로 나눠 떨어뜨릴 수 있어 최대잉여법이
  필요 없다. 대신 수수료만큼 예산을 미리 빼고 배분한다 — 온주 거래는
  내림에서 남는 돈이 수수료를 덮지만, 소수점 거래는 남는 돈이 0이라
  그냥 두면 잔여 현금이 음수가 된다. 1차 배분으로 수수료를 어림한 뒤
  줄어든 예산으로 다시 배분한다(수수료율이 0.015% 수준이라 한 번이면 수렴).
- 수량은 소수점 4자리에서 버림. 반올림하면 배분 합계가 총액을 넘는다.
  남는 금액은 `주가 × 0.0001` 수준이라 사실상 0이고, 6자리보다 읽기 좋다.
- 부동소수 찌꺼기(1e-5 미만)는 거래로 잡지 않는다.
- 허용 편차 안이라 잠긴 종목은 이제 현재 수량을 그대로 유지한다.
  기존 코드의 `.toInt()`는 3.45주를 3주로 깎아 없던 매도를 만들었다.

표시:
- utils/share_format.dart 에 수량 표기를 한곳으로 모았다. 계산에 쓴 값과
  화면에 찍히는 값이 같아야 브로커 앱에 그대로 옮겨 적을 수 있다.
- 상세 화면 리밸런싱 헤더 하단에 `· 소수점 거래`를 한 번만 표시.
- 거래 확인 다이얼로그의 수량 입력이 소수점을 받는다(온주일 땐 정수만).

설정 다이얼로그에 `소수점 거래` 항목 추가 (ko/en).

flutter analyze error 0 / warning 0, flutter test 73/73 (신규 16).

+++ test/rebalancer_threshold_test.dart
@@ -36,10 +36,10 @@ void main() {
       );
 
   /// id → delta 로 정리
-  Map<String, int> deltas(RebalanceResult r) =>
+  Map<String, double> deltas(RebalanceResult r) =>
       {for (final x in r.results) x.id: x.delta};
 
-  Map<String, int> shares(RebalanceResult r) =>
+  Map<String, double> shares(RebalanceResult r) =>
       {for (final x in r.results) x.id: x.newShares};
 
   // a 52%, b 48% / 목표 50:50 → 편차 ±2%p
